# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de la plateforme QUI PERD (défis de match, escrow, paiements Mobile Money).",
        "title": "QUI PERD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/administration/configurations-financieres": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "administration"
                ],
                "summary": "Configurations financières actives (admin)",
                "responses": {}
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "administration"
                ],
                "summary": "Modifier une configuration financière (admin) — historisé",
                "responses": {}
            }
        },
        "/administration/journaux-audit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "administration"
                ],
                "summary": "Journal d'audit (admin, lecture seule) — paginé, 10 par page, plus récents d'abord",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (défaut 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Éléments par page (1..100, défaut 10)",
                        "name": "taille",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtre exact sur l'action (ex. litige:decision)",
                        "name": "action",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Page-administration_JournalAudit"
                        }
                    }
                }
            }
        },
        "/administration/statistiques": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "administration"
                ],
                "summary": "KPIs du tableau de bord admin (cache Redis 60 s)",
                "responses": {}
            }
        },
...
        "/auth/changer-mot-de-passe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Changer son mot de passe",
                "responses": {}
            }
        },
...
        "/auth/deconnexion": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Déconnexion (invalide la session Redis)",
                "responses": {}
            }
        },
        "/auth/inscription": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Créer un compte joueur",
                "parameters": [
                    {
                        "description": "Compte à créer",
                        "name": "corps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.EntreeInscription"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/moi": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Utilisateur ou administrateur courant",
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "administration.JournalAudit": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "administrateurId": {
                    "type": "string"
                },
                "adresseIp": {
                    "type": "string"
                },
                "ancienneValeur": {
                    "type": "string"
                },
                "dateCreation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identifiantCible": {
                    "type": "string"
                },
                "nouvelleValeur": {
                    "type": "string"
                },
                "tableCible": {
                    "type": "string"
                },
                "utilisateurId": {
                    "type": "string"
                }
            }
        },
        "auth.EntreeInscription": {
            "type": "object",
            "required": [
                "email",
                "motDePasse",
                "nomUtilisateur"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "motDePasse": {
                    "type": "string",
                    "minLength": 6
                },
                "nomUtilisateur": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "pays": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
        "utils.Page-administration_JournalAudit": {
            "type": "object",
            "properties": {
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/administration.JournalAudit"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "taille": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}